+++ src/main.py
@@ -1,34 +1,52 @@
-import json
+import time
 import random
-from random import random
+import threading
+import contextlib
 from typing import List, Dict
 from os import environ as env
 
+import uvicorn
+from fastapi import FastAPI
 from rich import print
 from dotenv import load_dotenv
 from pytwitter import StreamApi
 from rich.traceback import install
 from datetime import datetime, timedelta
 
-from db import connect_to_redis
-from helpers import reset_rules
+from helpers import is_valid_text, reset_rules
 from _types import Tweet, StoredObject
 
 load_dotenv()
 install()
 
-redis_client = connect_to_redis()
+app = FastAPI()
+
+class Server(uvicorn.Server):
+    def install_signal_handlers(self):
+        pass
+
+    @contextlib.contextmanager
+    def run_in_thread(self):
+        thread = threading.Thread(target=self.run)
+        thread.start()
+        try:
+            while not self.started:
+                time.sleep(1e-3)
+            yield
+        finally:
+            self.should_exit = True
+            thread.join()
 
 # If set to false, it will delete all the rules and use the RULES from src\data.py
-is_rule_ok = True
+is_rule_ok = False 
 
 last_updated = datetime.now()
 
 super_dict: Dict[str, List[StoredObject]] = {"meme_stream": [], "tech_stream": []}
+storage_dict : Dict[str, List[StoredObject]] = {"meme_stream": [], "tech_stream": []}
 
 stream = StreamApi(bearer_token=env.get("TWITTER_BEARER_TOKEN"))
 
-print("[green]Starting stream[/green]")
 
 def handle_tweet(tweet: Tweet):
 
@@ -43,15 +61,15 @@ def handle_tweet(tweet: Tweet):
 
     # If the created_at is less than one month ago, it will not be saved
     if created_at > datetime.now() - timedelta(days=30):
+        print("[red]User acc created less than a month ago, skipping[/red]")
         return
     if not "media" in tweet["includes"]:
+        print("[red]No media, skipping[/red]")
         return
     if not tweet["includes"]["media"][0]["type"] == "photo":
+        print("[red]Not a photo, skipping[/red]")
         return
 
-    # Check if tweet["includes"]["users"][0]["name"] contains english characters only
-    if not all(ord(c) < 128 for c in tweet["includes"]["users"][0]["name"]):
-        return
 
     stored_object: StoredObject = {
         "username": tweet["includes"]["users"][0]["username"],
@@ -66,40 +84,48 @@ def handle_tweet(tweet: Tweet):
 
     if tweet["matching_rules"][0]["tag"] == "Funny things":
         super_dict["meme_stream"].append(stored_object)
-    elif tweet["matching_rules"][0]["id"] == "1544187580954349569":
+    else:
         super_dict["tech_stream"].append(stored_object)
 
 
 def flush_and_reset():
 
     # delete meme_Stream
-    redis_client.delete("meme_stream")
-    print("[green]Flushed memes database[/green]")
-    redis_client.set("meme_stream", json.dumps(super_dict["meme_stream"]))
-    print("[green]Saved memes to redis[/green]")
+    storage_dict["meme_stream"] = super_dict["meme_stream"]
+    super_dict["meme_stream"] = []
 
-    if random.randint(0, 1) == 1:
-        # delete tech_Stream
-        redis_client.delete("tech_stream")
-        print("[green]Flushed tech database[/green]")
-        redis_client.set("tech_stream", json.dumps(super_dict["tech_stream"]))
-        print("[green]Saved tech to redis[/green]")
+    storage_dict["tech_stream"] = super_dict["tech_stream"]
+    super_dict["tech_stream"] = []
 
+    print(storage_dict)
 
 stream.on_tweet = handle_tweet
 
 if not is_rule_ok:
     reset_rules(stream)
 
-stream.search_stream(
-    tweet_fields=["created_at"],
-    user_fields=[
-        "username",
-        "name",
-        "profile_image_url",
-        "created_at",
-    ],  # To get the username
-    expansions=["attachments.media_keys", "author_id"],
-    media_fields=["preview_image_url", "url"],  # To get the image
-    return_json=True,  # Return JSON because pytwitter doesn't return the `includes` key
-)
+
+@app.get("/get_memes")
+async def get_memes():
+    """Get the current memes stored in cache"""
+    return storage_dict
+
+
+# Asynchrounosly start the server
+config = uvicorn.Config(app=app)
+server = Server(config)
+
+with server.run_in_thread():
+    print("[green]Starting stream[/green]")
+    stream.search_stream(
+        tweet_fields=["created_at"],
+        user_fields=[
+            "username",
+            "name",
+            "profile_image_url",
+            "created_at",
+        ],  # To get the username
+        expansions=["attachments.media_keys", "author_id"],
+        media_fields=["preview_image_url", "url"],  # To get the image
+        return_json=True,  # Return JSON because pytwitter doesn't return the `includes` key
+    )